Reimplement prefetching by storing a JSON file of a subset of the MRU list.
This replaces the old prefetching code which would attempt to load N most
recent {firefox,xul}.pdb files from the symbol directories, instead simply
saving the N most recently-used symbol files in a JSON state file and reloading
them on startup.

The prefetching timer which would look for new symbol files is not implemented,
as this is groundwork for switching symbol fetching to use HTTP where that
would not be feasible.