Move symbolication processing to its own process.
In the efforts to move from the current "one thread per connection"
model to reactor pattern, we process the CPU intensive symbolication
code to its own process, so requests don't get stuck waiting for it to
finish.

Notice the locks in symFileManager are not necessary anymore and we
removed them.

multiprocessing module and Tornado are not good friends, and because of
that we use concurrent.futures.