Implement on disk symbol file cache.
This commit adds support for on disk symbol file cache. The cache is
done through the pickle module. We maintain a list of the most recently
used symbol files. The MRU is shared between the memory and disk cache.
The first "maxEntries" entries in the MRU corresponds to the memory
cache.

We also refactor the symbol fetching code.

We apply some OOP to symbol fetching here. The concept of a symbol
fetcher object is a interface which will receive a lib and a breakpad id
and return the respective SymbolInfo object, or None if the SymbolInfo
object is not found.

There are currently two fetchers:
* PathFetcher: look for symbol files in local paths.
* URLFetcher: look for symbol files in a remote path.

The symbol file manager constructs a pipeline of fetchers, calling one
by one until either one of them returns a SymbolInfo or all fetchers
are called.